+++ src/vertex_array_object.rs
@@ -50,8 +50,7 @@ impl VertexAttributesSystem {
             }
             for vertex_buffer in vertex_buffers.iter() {
                 buffers_list.push(match vertex_buffer {
-                    &&VerticesSource::VertexBuffer(ref vb, _, _) => vb.get_id(),
-                    &&VerticesSource::PerInstanceBuffer(ref buf) => buf.get_id(),
+                    &&VerticesSource::VertexBuffer(ref vb, _, _, _) => vb.get_id(),
                 });
             }
             buffers_list.sort();
@@ -136,11 +135,8 @@ impl VertexArrayObject {
         // checking the attributes types
         for vertex_buffer in vertex_buffers.iter() {
             let bindings = match vertex_buffer {
-                &&VerticesSource::VertexBuffer(ref vertex_buffer, _, _) => {
+                &&VerticesSource::VertexBuffer(ref vertex_buffer, _, _, _) => {
                     vertex_buffer.get_bindings()
-                },
-                &&VerticesSource::PerInstanceBuffer(ref buffer) => {
-                    buffer.get_bindings()
                 },
             };
 
@@ -164,11 +160,8 @@ impl VertexArrayObject {
             let mut found = false;
             for vertex_buffer in vertex_buffers.iter() {
                 let bindings = match vertex_buffer {
-                    &&VerticesSource::VertexBuffer(ref vertex_buffer, _, _) => {
+                    &&VerticesSource::VertexBuffer(ref vertex_buffer, _, _, _) => {
                         vertex_buffer.get_bindings()
-                    },
-                    &&VerticesSource::PerInstanceBuffer(ref buffer) => {
-                        buffer.get_bindings()
                     },
                 };
 
@@ -187,20 +180,12 @@ impl VertexArrayObject {
         // building the values that will be sent to the other thread
         let data = vertex_buffers.iter().map(|vertex_buffer| {
             match vertex_buffer {
-                &&VerticesSource::VertexBuffer(ref vertex_buffer, _, _) => {
+                &&VerticesSource::VertexBuffer(ref vertex_buffer, _, _, per_instance) => {
                     (
                         GlObject::get_id(*vertex_buffer),
                         vertex_buffer.get_bindings().clone(),
                         vertex_buffer.get_elements_size(),
-                        0 as u32
-                    )
-                },
-                &&VerticesSource::PerInstanceBuffer(ref buffer) => {
-                    (
-                        GlObject::get_id(*buffer),
-                        buffer.get_bindings().clone(),
-                        buffer.get_elements_size(),
-                        1 as u32
+                        if per_instance { 1 as u32 } else { 0 as u32 }
                     )
                 },
             }